Use numpy dtypes for our dtypes.
Fixes #15.

Specifically, use an actual dtype object and not the type.
For example, `np.int64` and `np.longlong` both map to `dtype('int64')`.  We use the latter.
When converting to dtype, we can handle every dtype-like thing that numpy does such as 'int64'.

We still have a set of dtypes that metagraph nominally understands and supports,
and code is written such that we could easily make our own dtype objects.

This opens up the world of dtype-aware translations.
To see where things may get painful, we include more dtypes than we strictly need right now.
Conversely, there are also numpy dtypes that we don't yet have, so we'll see how conversions
between numpy-aware libraries work out.

Also, this makes numpy a hard dependency.  This seems reasonable to me.
I have not yet updated code that was conditional on numpy being installed,
because I'm afraid of merge conflicts.  We'll update this after merges.