Reduce number of channels in environment.yml
Changes in https://github.com/ContinuumIO/metagraph/pull/24/commits/c78b491b8c7b09a41447994aa9ac31a195e043b8 made it so that we added the channel jim22k to the environment.yml ; since the channel is only used for one dependency, i.e. grblas, it would be nicer to just inline the channel for the one case it is used in. This commit is intended to do this.