Handle typing.Any in abstract algorithm signatures.
Sometimes, e.g. with BFS, we might want to specify a node as an input.

We don't know the type when declaring the abstract algorithm signature.

Different graphs store nodes in different ways. They could be integers, strings, etc.

Thus, we should declare them more specifically at concrete algorithm declaration time.

Previously, when we declared that the type of a parameter was typing.Any in the abstract algorithm signature, it wouldn't match with anything besides typing.Any in the concrete algorithm signature.

That's been fixed so that if typing.Any is the declared type on an abstract algorithm parameter, it'll match anything in the concrete algorithm signature.