Support Tuple Return Types
We can now in our abstract and concrete algorithms declare the return type to be a tuple.

This required support to be added in:
* Resolver._normalize_concrete_algorithm_signature, where concrete algorithm signature checking happens.
* MultiVerify.assert_equals, which is used to compare values during testing.

Misc. Notes:
* The logic of the outer conditional (that determined which path to take based on the concrete return type) was flipped in the implementation of _check_concrete_algorithm_return_signature from that of _normalize_concrete_algorithm_signature for both readability and commonality of the 3 cases.

This allows us to commit the test, concrete algorithm, and abstract algorithm declaration for Bellman-Ford in the NetworkX plugin. This commit includes these changes.