Move logic for tuple support from MultiVerify._compare_values -> MultiVerify.assert_equals
It was suggested in https://github.com/ContinuumIO/metagraph/pull/35#pullrequestreview-402004584 to move the support for tuples from MultiVerify._compare_values to MultiVerify.assert_equals.

This suggestion makes a lot of sense as it simplifies the functionality of MultiVerify._compare_values (and thus makes its intended purpose more understandable) and reduces the amount of recursion, which tends to be unreadable. This does remove support for nested tuples, but it's not clear that we desire support for that case currently since it is so rare.