Move logic for tuple support from Resolver._check_concrete_algorithm_return_signature -> Resolver._normalize_concrete_algorithm_signature
A previous commit for the testing harness changes moved tuple checking logic from MultiVerify._compare_values -> MultiVerify.assert_equals.

This patch makes the analogous change in the Resolver class.

Without this change, we'd be able to write algorithms returning nested tuples but not test them. This change makes it so the tuple return signatures for abstract and concrete algorithms are strictly the ones that can be tested by MultiVerify.