Refactor plugin regsitration
This commit makes it so that the find_plugins functions used in entry point declarations for the purpose of delcaring plugins now returns a nested dictionary that has the topology (plugin_name -> plugin_attribute_name -> attribute_set).

This will make it so that we can easily define a find_plugins function by just inlining a nested dictionary. We can still use a PluginRegistry if desired by having the find_plugins function return the result of PluginRegistry.plugins .

This commit also makes it so that instances of PluginRegistry have a default plugin name. We can explicitly specify a plugin name when calling PluginRegistry.register_from_modules or use the default.

This commit makes it so that metagraph.plugins.find_plugins registers the core abstract types and algorithms under the plugin name "core" and the concrete types, wrappers, etc. for each core library (e.g. NetworkX) under the names "core_*" (e.g. "core_networkx").

Here are more misc. details:
* metagraph.core.entrypoints.load_plugins now returns a dictionary of plugins (plugin_name -> plugin_attribute_name -> attribute_set)
* This commit recombines PluginRegistry.register_abstract and PluginRegistry.register_concrete into PluginRegistry.register since we can now associate abstract algorithms with plugins (no need to have two more specific methods).
* The Resolver class has been refactored to account for the above changes. A lof of the changes to tests are not very interesting.
* Tests were refactored to match the above changes. A lof of the changes to tests are not very interesting.