Make plugin registration with resolver code cleaner
This commit inteds to make the registration of abstract types, wrappers, etc. functionality in the Resolver.register method more compact. This is intended to be a no-op refactoring for redability. This commit does make it so that Resolver._register_plugin_attributes_in_tree takes optional arguments now, so it's not 100% no-op.

This commit also gets rid of example_plugin_util.py and moves it back into util.py .

This comit also gets rid of plugin1_util.py and moves it back into plugin1.py .