New concept of unambiguous_subcomponents
This allows passing an EdgeMap to an algorithm which needs an EdgeSet.
Normally translators can only convert between concrete versions of the
same abstract type, but unambiguous_subcomponents tells the system
that additional translators are allowed because they can be called without
requiring any arguments and will give an unambiguous answer.