Complete first draft of Plugin Author Guide.
This commit completes the first draft of the Plugin Author Guide.

We completed:
* Plugin Parts
* End-to-End Plugin Pathway
* Plugin Testing

We included links to some external pages.

Each of the docs contain a "Prerequisite Reading" section to specifically enumerate what's necessary to know before reading the presented document.

We went over practical ways of thinking about plugins in "Plugin Parts".

We included concrete examples of a plugin in "End-to-End Plugin Pathway".

We went over how to use MultiVerify in "Plugin Testing".