Complete first draft of Algorithm List and Visualization parts of the User Guide.
This commit completes the first draft of the Algorithm List and Visualization sections of the User Guide.

The algorithm list contains a description of all the abstract algorithms currently supported in the metagraph core plugins.

The visualization section describes why visualization's difficult to fit within the metagraoh workflow and forwards readers to suggestion any ideas they might have.