Fix weights used in air travel use case tutorial
The air travel use case notebook used the number of passengers as weights in the graph used to find the betweenness centrality (which we intended to correspond to the most travelled through areas).

This is incorrect since paths with many travellers would have a larger total weight. This is the opposite of what's desired since we want paths with more travellers to have a smaller total weight (i.e. be the shorteer paths) so that the betweenness centrality can denote the areas that more travellers pass through.

This commit fixes this issue in the notebook by using 1/passenger_count as the weight. A more elegant metric can be used, but we chose a simple one to make the tutorial easier to consume.