Pass node lists to returned values of ss_all_shortest_lengths
If the graph passed into ss_all_shortest_lengths has a node list that isn't an identity mapping (which is a lot of the time), we'll get incorrect results returned from ss_all_shortest_lengths as the generated edge maps will have the wrong labels.

This commit passes down the node_list from the input graph to the returned edge maps.