Make examples flow more sanely in "End-to-End Plugin Pathway"
It was pointed out that the explanation of the plugin-finder function return type was unclear since it used an example that deviated away from the working example. This commit adds some more verbose explanation in an attempt to alleviate this issue. This is a progress patch. There's more progress to come.