Reorganize "End-to-End Plugin Pathway" in a more easily-consumable manner
It was pointed out that the ordering of concepts in the End-to-End Plugin Pathway tutorial was confusing.

This commit reorders the tutorial.

We first go over how to use a plugin registry to write a plugin-finder function before going over what exactly the return type of a plugin-finder function looks like.