Fix type coercion problem in np_nodemap_apply
https://github.com/ContinuumIO/metagraph/pull/91#pullrequestreview-453615924 reported an issue with the previous implementation of np_nodemap_apply.

Previously, we were copying the original array, applying the function, and inserting those values back into the copy.

This is problematic because the function might return something whose type is different from the dtype of the original array's copy.

This commit fixes that by first getting the results of applying the function and then creating the array using the dtype of the results.