Document behavior of util.graph.aggregate_edges and make ss_graph_filter_edges not use the dangerous eliminate_zeros method
We used to use spmatrix.eliminate_zeros in ss_graph_filter_edges, which would cause problems if there are edges with zero weight. This commit implements util.graph.filter_edges for the Scipy plugin that doesn't rely on this method.

This commit also more clearly documents util.graph.aggregate_edges and it's parameters.