Remove notebooks from metagraph/docs/_downloads/
The metagraph/docs/_downloads/ directory used to contain compressed files with the raw data and notebooks.

Since we can now download the notebooks directly from the generated docs and since the notebooks link to the official source of the data, the _downloads directory is no longer needed. This commit removes them.