PlanNamespace.translate to return a MultiStepTranslator
This commit makes PlanNamespace.translate return a MultiStepTranslator always. If no translation path is possible, a MultiStepTranslator is still returned with the attribute of unsatisfiable set to True.

The tests have been updated accordingly.

When trying to get the length or iterate through a MultiStepTranslator that is unsatisfiable, a ValueError is raised.

This commit also makes it so that the __repr__ method of MultiStepTranslator returns the same thing that MultiStepTranslator.display prints out.