Add attribute AlgorithmPlan.build_problem_messages
This commit adds the attribute AlgorithmPlan.build_problem_messages, which will tell us if the algorithm plan is valid.

This also makes the class method AlgorithmPlan.build always return something instead of retuning None when a problem occurs. If the returned plan is problematic, there will be a complaint message in AlgorithmPlan.build_problem_messages .