Make MultiStepTranslator error meassages robust against empty dst_types attribute
The MultiStepTranslator used to raise exceptions with the message f"No translation path found for {src_type.__name__} -> {self.dst_types[-1].__name__}" when a translation couldn't be satisfied.

This would break when the dst_types was empty.

This commit makes it so that instead of storing a bool named unsatisfiable, we store the destination type that the MultiStepTranslator is not able to translate to in the attribute unsatisfiable_dst_type .