Add final_type attribute to MultiStepTranslator
This commit adds the attribute final_type to MultiStepTranslator.

This commit also makes it so that the contructor for MultiStepTranslator takes in an additional final_type parameter.

This allows us to replace the unsatisfiable_dst_type attribute with the attributes final_type and unsatisfiable (also added in this patch). This makes the interface to MultiStepTranslator easier to comprehend.