Move out hard-coded HTML, CSS, and JS of metagraph/metagraph/explorer/service.py to template files
This commit is intended to be a no-op.

This commit is a reorganization of metagraph/metagraph/explorer/service.py .

It moves out the hard-coded HTML, CSS, and JS out to template files so that they are easier to manage (and read because of extension-specific syntax highlighting).

This commit also moves the shadow element's script tag JS code inside of a larger encapsulating anonymous function in order to limit the scoping of variables. This allows us to have fewer string replacements done on the Python-side, which will lead to more maintainability and cleaner code.