Translation Graph Explorer: Stop simulations between runs
Previously, we used to create a new simulation everytime we changed the abstract type in the translation graph explorer.

This would leave the old simulation running (memory leak + errors in the old one would lead to the translation graph explorer breaking).

We now stop the old simulation when changing abstract types.