flow.max_flow concrete algorithms for Scipy & NetworkX Plugins
This commit adds the max flow algorithms for the Scipy and NetworkX plugins.

NOTE: For the Scipy plugin implementation, the result of scipy.sparse.csgraph.minimum_spanning_tree (stored in span_tree) is not guaranteed to be symmetric or to be such that M[i][j] != M[j][i] for non-zero values where M is the matrix. Thus, the Scipy plugin implementation DWIMs the result to be symmetric via (M + M.T) / ( bool(M) + bool(M.T) ) where bool takes a matrix and returns the equivalent boolean mask and the division only happens for non-zero values.