Add NetworkX and Scipy Plugin implementations for max flow
This commit adds the NetworkX and Scipy Plugin implementations for max flow.

This commit also changes MultiVerify. It abstracts out the translation functionality. This is intended to make the code easier to read since it'll reduce the number of try/except blocks.

There's a test for max flow as well. We may want to consider adding more test coverage since max flow doesn't necessarily have a unique solution for a given graph.

The Scipy plugin implementation also only works for int weights. An exception is raised if float weights are given (which isn't handled right now).