Explorer: Close the explorer with an error message if there's a websocket error.
When exceptions are raised on the Python-side of the metargaph explorer, the explorer used to die silently while the websocket became closed.

This commit makes it so that if the explorer tries to send a websocket request and the websocket is closed, the explorer closes and displays an error message.