Explorer: Remove Python-side filtering
We used to have a "filters" parameter in the Python-side websocket API so that the UI could display only certain edges or types.

It's much easier to do this on the JS-side as it'll only require CSS class changes and won't require us to re-render the translation graph (which currently happens upon receiving a websocket response, which is necessary if the filtering is done on the Python-side).

This commit removes the filtering that's done on the Python-side.