Algorithm Type Solver Explorer: Show current signature in algorithm type solver
This commit makes it so that we show the current algorithm signature (with all the concrete types) when we change the selections for the concrete type dropdowns.

This is so that the users can see that the dropdown actually made a change since very frequently solving the algorithm types yields the same number of plans, which would yield no visual change that would indicate that anything happened.