Algorithm Type Solver Explorer: Check if plugin has concrete algorithm before accessing it.
The explorer API function solve_algorithm used to access resolver.SOME_PLUGIN.concrete_algorithms[SOME_ABSTRACT_ALGO] , which would cause resolver.SOME_PLUGIN.concrete_algorithms (which is a defaultdict) to create an entry for SOME_ABSTRACT_ALGO even if it didn't exist before. This means the explorer would change the state of the resolver, which is undesired. This commit removes this access.