Documentation: Update Use Case Tutorial 3 to use metargaph.NodeLabels
The ecommerce tutorial has a bipartite graph of customer ids (ints) and stock codes (strings).

Some backends require all nodes of a bipartite graph to all be instances of the same type.

It's best practice to use metagraph.NodeLabels to have the nodes of graphs be the same type.

This commit makes the ecommerce tutorial use metagraph.NodeLabels.