Add node embedding type
This commit adds the following abstract algorithms:
* embedding.train.node2vec
* util.node_embedding.apply

This commit adds the abstract type NodeEmbedding and the wrapper NodeEmbeddingWrapper.

The motivation for this is:
* A transductive embedding can only embed things seen during training.
* An inductive can embed things unseen during training.
* Looking at the embedding algorithms we currently intend to implement (graph2vec, node2vec, HOPE,  LINE, graph_sage, graph_wave), there are 3 types of embedding abstract types we need. They are the transductive node embedding type, transductive graph embedding type, and inductive node embedding type. Our current list of embedding algorithms don't motivate the need for an inductive graph embedding type.
* Applying a transductive node embedding to a node will require only a node (since the graph is fixed and was seen during training) and will return a vector.
* Applying a transductive graph embedding to a graph will require only a graph (that was seen during training) and will return a vector.
* Applying an inductive node embedding will require a graph and a node (neither of which were necessarily seen during training) and will return a vector corresponding to the node.
* The node2vec, HOPE, LINE, and graph_wave algorithms produce a transductive node embedding.
* The graph2vec algorithm produces a transductive graph embedding.
* The graph_sage algorithm produces an inductive node embedding.

embedding.train.node2vec is the abstract algorithm to perform the training phase of the node2vec algorithm to produce a transductive node embedding (in the form of a concrete type instance corresponding to NodeEmbedding).

util.node_embedding.apply will take a transductive node embedding and vector of node ids and will return a matrix of the embedding vectors (each row corresponds to the embedding vector of the given node ids; the order of thw rows match the order of the input vector of node ids).

Perhaps we'll want to change the naming convention to call it "util.transductive_node_embedding.apply" to differentiate it from the not-yet-implemented inductive node embedding abstract algorithm. We'll revisit this when we implement the inductive node embedding abstract algorithm.

This commit contains a test for util.node_embedding.apply (see test_node_embedding_apply).

This commit adds the concrete type NumpyNodeEmbedding.

This commit also adds array accessing for NumpyNodeMap, e.g.

    nm = NumpyNodeMap(np.array([100, 101, 99999, 103]), mask=np.array([1, 1, 0, 1],  dtype=bool))
    selector = np.array([0,3])
    nm[selector] # this is now valid

This only works for np.ndarray instances (but not Python lists as is supported in numpy; we could easily add this if desired).

There's a test for this in test_python_2_numpy_mask.

This commit also adds the method np_matrix for NumpyNodeMap. It simply returns the numpy matrix (accounting for the mask if necessary).