Make NumpyMatrix.as_dense guarantee no copy is made when copy=False
Previously, NumpyMatrix.as_dense used to make a copy of NumpyMatrix.value when NumpyMatrix.mask was not None even if copy=False was passed in.

It now does not force a copy when NumpyMatrix.mask is not None.

This does change NumpyMatrix.value in place when NumpyMatrix.as_dense is called. This can be dangerous as other references to previously returned values from NumpyMatrix.as_dense(copy=False) are used as the refenced value will be changed, but that's to be expected when copy=False is specified.