Make transductive node embeddings return Tuple[Matrix, NodeMap]
This commit makes it so that the following algorithms return a Tuple[Matrix, NodeMap] .
* embedding.train.node2vec
* embedding.train.graphwave
* embedding.train.hope.katz

This commit also gets rid of the NodeEmbedding abstract type and concrete type in the numpy plugin.

Because of this, the signature for embedding.train.graph_sage.mean had to change as well.

Similarly, the signatures for util.node_embedding.apply and util.graph_sage_node_embedding.apply changed.

The tests were updated to account for the above changes.