Add mg.List and graph2vec abstract algorithm.
This commit adds mg.List, which takes exactly one argument. It denotes a python list containing elements of exactly one type.

This commit also adds the graph2vec abstract algorithm.

This commit also contains a test for graph2vec. It uses sklearn. The test is skipped if sklearn is not installed.

This commit also contains a test for mg.List in the new test test_list_signatures.