Change PythonNodeSet/Map wrappers to be concrete types
PythonNodeSet is now just a set with PythonNodeSetType concrete type
PythonNodeMap is now just a dict with PythonNodeMapType concrete type

Fingerprints were needed for the typecache to handle caching properties
for dict which is unhashable and cannot be a weakref. This lets us check
 the fingerprint of each dict to see if the id is being reused or not
and therefore invalidate the property cache.