Reorder resolver's type checking methods to follow a top-down approach
Since reading the code for the resolver's type checking functionality can be confusing given the large size of the code, it would help if the methods were ordered in a way that mirrored the control flow graph / call graph.

This commit aims to reorder the methods called directly or indirectly by _register_plugin_attributes_in_tree in a top-down sort of way where methods only called directly or indirectly by _register_plugin_attributes_in_tree are defined after we see an instance of their use.

The goal is to make it so that reading the file from top-to-bottom goes from abstract to less abstract rather than having the more or less jumbled up ordering we had before where when reading the file from top-to-bottom, we couldn't tell at what level of abstraction the current method would be.

This is reordering aims to make it so that reading the code/methods from the top of the file to the bottom of the file follows a topological sort of the call graph (with the additional constraint that direct neighbors are as close as possible, i.e. the topological sort is "greedy") with as the root node _register_plugin_attributes_in_tree.