Default resolver to enable wrapper methods
Wrapper methods include `run` and `translate`. They are functionally equivalent
to calling `resolver.run` and `resolver.translate` with the wrapper as the first argument.
The `dst_type` of the translate call has been expanded to allow more possible ways to
indicate the destination type, including instances of the wrapper and strings of the class names.